Highlighting and Live Search
* Fixed mainWindow not storing the find and replace window. Thus not
allowing use of commands for it when not focused on the window.
* Added live searching.
* Added search result highlighting.
* Switching between find/replace through hotkeys now properly initiates
focus on the find textbox after the change.
