+++ ZeldaOracle/ConscriptDesigner/Windows/FindReplaceWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="ConscriptDesigner.Windows.FindAndReplaceWindow"
+﻿<Window x:Class="ConscriptDesigner.Windows.FindReplaceWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -32,7 +32,7 @@
             <TabItem Header="Find" IsTabStop="False">
                 <StackPanel>
                     <TextBlock Margin="3">Text to Find:</TextBlock>
-                    <TextBox Margin="3" Name="textBoxFind" />
+                    <TextBox Margin="3" Name="textBoxFind" TextChanged="OnFindTextChanged" />
                     <StackPanel Orientation="Horizontal" HorizontalAlignment="Right">
                         <Button Margin="5" HorizontalAlignment="Right" Width="80" Content="Find Next" Click="OnFindNext" IsDefault="True" />
                     </StackPanel>
@@ -41,23 +41,24 @@
             <TabItem Header="Replace" IsTabStop="False">
                 <StackPanel>
                     <TextBlock Margin="3">Text to Find:</TextBlock>
-                    <TextBox Margin="3" Name="textBoxReplaceFind" Text="{Binding ElementName=textBoxFind, Path=Text}"/>
+                    <TextBox Margin="3" Name="textBoxReplaceFind" Text="{Binding ElementName=textBoxFind, Path=Text, Mode=TwoWay}" TextChanged="OnFindTextChanged"/>
                     <TextBlock Margin="3" Text="Replace with:" />
                     <TextBox Margin="3" Name="textBoxReplace" />
                     <StackPanel Orientation="Horizontal" HorizontalAlignment="Right">
-                        <Button Margin="5" HorizontalAlignment="Right" Width="80" Content="Find Next" Click="OnReplaceFindNext" IsDefault="True" />
-                        <Button Margin="5" HorizontalAlignment="Right" Width="80" Content="Replace" Click="OnReplace" />
+                        <Button Margin="5" HorizontalAlignment="Right" Width="80" Content="Find Next" Click="OnFindNext" IsDefault="True" />
+                        <Button Margin="5" HorizontalAlignment="Right" Width="80" Content="Replace" Click="OnReplaceNext" />
                         <Button Margin="5" HorizontalAlignment="Right" Width="80" Content="Replace All" Click="OnReplaceAll" />
                     </StackPanel>
                 </StackPanel>
             </TabItem>
         </TabControl>
 
-        <CheckBox Grid.Row="2" Grid.Column="0" Margin="10,2" Name="checkBoxCaseSensitive" Content="Match case" IsTabStop="False" />
-        <CheckBox Grid.Row="3" Grid.Column="0" Margin="10,2" Name="checkBoxWholeWord" Content="Match whole word" IsTabStop="False" />
-        <CheckBox Grid.Row="2" Grid.Column="1" Margin="10,2" Name="checkBoxRegex" Content="Regular Expression" IsChecked="false" IsTabStop="False" />
-        <CheckBox Grid.Row="3" Grid.Column="1" Margin="10,2" Name="checkBoxWrapSearch" Content="Wrap search" IsChecked="false" IsTabStop="False" />
-        <CheckBox Grid.Row="4" Grid.Column="1" Margin="10,2" Name="checkBoxSearchUp" Content="Search up" IsChecked="false" IsTabStop="False" />
-        <ComboBox x:Name="comboBoxScope" Grid.Row="5" Grid.ColumnSpan="2" Margin="10,2" IsTabStop="False"/>
+        <CheckBox Grid.Row="2" Grid.Column="0" Margin="10,2" Name="checkBoxCaseSensitive" Content="Match case" IsTabStop="False" Click="OnCheckboxChanged" />
+        <CheckBox Grid.Row="3" Grid.Column="0" Margin="10,2" Name="checkBoxWholeWord" Content="Match whole word" IsTabStop="False" Click="OnCheckboxChanged" />
+        <CheckBox Grid.Row="2" Grid.Column="1" Margin="10,2" Name="checkBoxRegex" Content="Regular Expression" IsChecked="false" IsTabStop="False" Click="OnCheckboxChanged" />
+        <CheckBox Grid.Row="3" Grid.Column="1" Margin="10,2" Name="checkBoxWrapSearch" Content="Wrap search" IsChecked="True" IsTabStop="False" Visibility="Hidden" />
+        <CheckBox Grid.Row="3" Grid.Column="1" Margin="10,2" Name="checkBoxSearchUp" Content="Search up" IsChecked="false" IsTabStop="False" />
+        <CheckBox Grid.Row="4" Grid.Column="1" Margin="10,2" Name="checkBoxLiveSearch" Content="Live search" IsChecked="false" IsTabStop="False" />
+        <ComboBox x:Name="comboBoxScope" Grid.Row="5" Grid.ColumnSpan="2" Margin="10,2" IsTabStop="False" SelectionChanged="OnScopeChanged"/>
     </Grid>
 </Window>
