+++ ZeldaOracle/ConscriptDesigner/Windows/SaveChangesWindow.xaml.cs
@@ -18,8 +18,20 @@ namespace ConscriptDesigner.Windows {
 	/// Interaction logic for SaveChangesWindow.xaml
 	/// </summary>
 	public partial class SaveChangesWindow : Window {
+
+		//-----------------------------------------------------------------------------
+		// Members
+		//-----------------------------------------------------------------------------
+
+		/// <summary>The returned message box result.</summary>
 		private MessageBoxResult result;
 
+
+		//-----------------------------------------------------------------------------
+		// Constructor
+		//-----------------------------------------------------------------------------
+
+		/// <summary>Constructs the save changes window.</summary>
 		private SaveChangesWindow(IEnumerable<string> files, bool reload) {
 			InitializeComponent();
 			foreach (var file in files) {
@@ -34,12 +46,23 @@ namespace ConscriptDesigner.Windows {
 			}
 		}
 
+
+		//-----------------------------------------------------------------------------
+		// Event Handlers
+		//-----------------------------------------------------------------------------
+
 		private void OnButtonClicked(object sender, RoutedEventArgs e) {
 			result = (MessageBoxResult)Enum.Parse(typeof(MessageBoxResult), ((Button) sender).Tag as string);
 			DialogResult = true;
 			Close();
 		}
 
+
+		//-----------------------------------------------------------------------------
+		// Showing
+		//-----------------------------------------------------------------------------
+
+		/// <summary>Shows the save changes window.</summary>
 		public static MessageBoxResult Show(Window owner, IEnumerable<string> files, bool reload) {
 			SaveChangesWindow window = new SaveChangesWindow(files, reload);
 			window.Owner = owner;
